requirements.txt
numpy
matplotlib
pyaudio
python-sounddevice
cuda-python
pyqtgraph